// src/app/shared/model/credit.ts
export class Credit {
    idCase: number;
    fechaLimiteAutogestion: Date;
    fechaDesembolso: string;
    obligacion: string;
    montoDesembolso: number;
    identificacionCliente: string;
    nombreCliente: string;
    tipoDocumento: string;
    numeroDocumento: string;
    valorDestino: number;
    descripcionRubro: string;
    codigoRubro: string;
    tipoRubro: string;
    description: string;
    groupCode: number;
    headingCode: string;
    amount: string;
    id: number;
    guidRubro: string;
    rubroCompleto: boolean;
    mensajeValidacionNotificacion: string;
    ajustesRechazo: string;
    fechaRechazo: string;
    justificacionRechazo: string;
    usuarioRechazo: string;
    state: string;
    obligationUpdate : boolean;
}